Refactor functions and dictionary out to a utils
