+++ src/App.js
@@ -3,15 +3,13 @@ import "./App.css";
 import { Heading } from "./components/Heading";
 import { HeadingGreeting } from "./components/HeadingGreeting";
 import { Timer } from "./components/Timer";
+import { getPayDateChecker } from "./utils/get-pay-date-checker";
 import moment from "moment";
 
 const App = () => {
   const { month, payDate } = Heading();
   const formattedDate = moment(payDate).format("dddd DD MMM").toUpperCase();
-  const dateChecker = () => {
-    const todaysDate = moment().format("L");
-    return todaysDate === payDate ? "" : "a countdown to";
-  };
+
   return (
     <div className="min-h-screen bg-white dark:bg-gray-900 ">
       <div className="container bg-white dark:bg-gray-900 pt-40 md:pt-0">
@@ -21,7 +19,8 @@ const App = () => {
           </h1>
           <Timer />
           <h1 className="mt-6 text-black dark:text-white" id="header">
-            {HeadingGreeting()}, this is {dateChecker()} your {month} Payday! 💰
+            {HeadingGreeting()}, this is {getPayDateChecker(payDate)} your{" "}
+            {month} Payday! 💰
           </h1>
         </div>
       </div>
