+++ src/utils/get-pay-date-checker.js
@@ -1,0 +1,6 @@
+import moment from "moment";
+
+export const getPayDateChecker = (payDate) => {
+  const todaysDate = moment().format("L");
+  return todaysDate === payDate ? "" : "a countdown to";
+};
