+++ src/utils/pay-date-dictionary.js
@@ -1,0 +1,16 @@
+const currentYear = new Date().getFullYear();
+
+export const payDateDictionary = [
+  { month: "January", payDate: `01/27/${currentYear}`, time: "00:00:00" },
+  { month: "February", payDate: `02/24/${currentYear}`, time: "00:00:00" },
+  { month: "March", payDate: `03/31/${currentYear}`, time: "00:00:00" },
+  { month: "April", payDate: `04/28/${currentYear}`, time: "00:00:00" },
+  { month: "May", payDate: `05/26/${currentYear}`, time: "00:00:00" },
+  { month: "June", payDate: `06/30/${currentYear}`, time: "00:00:00" },
+  { month: "July", payDate: `07/28/${currentYear}`, time: "00:00:00" },
+  { month: "August", payDate: `08/25/${currentYear}`, time: "00:00:00" },
+  { month: "September", payDate: `09/29/${currentYear}`, time: "00:00:00" },
+  { month: "October", payDate: `10/27/${currentYear}`, time: "00:00:00" },
+  { month: "November", payDate: `11/30/${currentYear}`, time: "00:00:00" },
+  { month: "Christmas", payDate: `12/22/${currentYear}`, time: "00:00:00" },
+];
